requirements-test.txt
# This requirements file should only include dependencies for testing
pytest                                  # test framework
pytest-cov                              # coverage test
lz4
python-snappy
pyzstd
